// src/app/commonBussinessSRC/psComponents/ps-lookup/ps-lookup-chequebooks/ps-lookup-chequebooks.component.interfaces.ts
import { Type } from '@angular/core';
import { IOptionsPsBaseFieldExposed } from 'src/app/commonSRC/psServices/models/ps-common-interface';

export interface IOptionsPsLookupChequebooksExposed extends IOptionsPsBaseFieldExposed {
    listOfBenificiaries?: any[];
    currency?: string;
    component?: Type<any>;
    showOnlyList?: boolean;
    type?: string;
    requestObject: any;
}